requirements.txt
gtts
pillow
webcolors